splitting out errors
