+++ src/parsers/mod.rs
@@ -8,6 +8,7 @@ pub mod row_parser;
 pub mod sanitiser;
 pub mod state;
 pub mod strategies;
+pub mod strategy_errors;
 pub mod strategy_file;
 pub mod strategy_structs;
 pub mod transformer;
